// src/app/features/organization/models/organization.enums.ts
export enum OrganizationType {
        CORPORATE = 'Corporate',
        CHURCH = 'Church',
        SCHOOL = 'School',
        NGO = 'NGO'
}

export enum ServiceType {
    EMPLOYEE_TRAINING = 'Employee Training & Development',
    COMPLIANCE_CERTIFICATION = 'Compliance & Certification Programs',
    FAITH_BASED_LEARNING = 'Faith-Based Learning & Leadership',
    CLASSROOM_MANAGEMENT = 'Classroom Management System',
    STUDENT_PROGRESS_TRACKING ='Student Progress & Analytics',
    LEARNING_MATERIAL_REPOSITORY = 'Leaning Material Repository',
    MEMBER_ENGAGEMENT = 'Member Engagement & Communication',
    CUSTOM_COURSE_CREATION = 'Custom Course Creation',
    LIVE_WORKSHOPS = 'Live Workshops & Webinars',
    FULL_PLATFORM_ACCESS = 'Full Platform Access'
 }

 export enum ServiceDuration {
    ONE_MONTH = 1,
    THREE_MONTHS = 3,
    SIX_MONTHS = 6,
    ONE_YEAR = 12,
    TWO_YEARS = 24,
    THREE_YEARS = 36,
    FOUR_YEARS = 48,
    FIVE_YEARS = 60,
    FIVE_PLUS = 61
 }

 export enum OrganizationStatus {
    ACTIVE = 'Active',
    INACTIVE = 'Inactive',
    PENDING = 'Pending',
    SUSPENDED = 'Suspended'
 }